requirements.txt
tensorflow>=2.15,<3.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
flask>=3.0
pillow>=10.0
